[VM] Move integration tests to own folder
