+++ machines/tests/integration/integration/test_exporter.py
@@ -10,7 +10,7 @@ import requests
 import tenacity
 from pytest_operator.plugin import OpsTest
 
-from .helpers import (
+from ..helpers import (
     APPLICATION_DEFAULT_APP_NAME,
     MYSQL_DEFAULT_APP_NAME,
     MYSQL_ROUTER_DEFAULT_APP_NAME,
