+++ machines/tests/integration/integration/test_exporter_with_tls.py
@@ -10,8 +10,8 @@ import requests
 import tenacity
 from pytest_operator.plugin import OpsTest
 
-from . import architecture, juju_
-from .helpers import (
+from .. import architecture, juju_
+from ..helpers import (
     APPLICATION_DEFAULT_APP_NAME,
     MYSQL_DEFAULT_APP_NAME,
     MYSQL_ROUTER_DEFAULT_APP_NAME,
