+++ machines/tests/integration/integration/test_hacluster.py
@@ -10,8 +10,8 @@ import pytest
 import tenacity
 from pytest_operator.plugin import OpsTest
 
-from . import architecture, juju_
-from .helpers import (
+from .. import architecture, juju_
+from ..helpers import (
     MYSQL_DEFAULT_APP_NAME,
     MYSQL_ROUTER_DEFAULT_APP_NAME,
     execute_queries_against_unit,
